build: replace contents of `prepare` script with Gulp task in package.json
The `prepare` script in package.json is used to build jest-chance fork (necessary because the
original jest-chance package does not allow Node.js version < 19.0.0). Using Gulp to to build
jest-chance provides better compatibility with Windows and gets rid of the `lefthook.yml` it
generates every time it is built.
NOTE: This `prepare` script for jest-chance is supposed to be temporary