build(gulp): fix pre-commit hook task
Fixes the issue where an error in any of the sub-tasks of the `precommitHook` task did not result in
the `precommitHook` task terminating in an error state. With this fix the pre-commit hook will fail
if the linter fails or any of the tests fail, like it is supposed to.