perf(home): move "start" button to top in hero section
On small devices the "start" button was cut off because of the headings in the hero section took up
a lot of space. The user would have to scroll down to get to the button. This situation is not ideal
because the button should be in a position that allows the user to quickly click on it as soon as
the page is loaded. Making the headings smaller and preserving the positioning would not be an ideal
solution because the headings could still take too much room in certain languages. Putting the
button at the top, although a bit awkward, makes the button immediately accessible for just about
every language and allows for the headings to remain the same size.