refactor: add workaround for unit tests failing in CI environment
The unit tests for pages in the `txn/` directory failed because of `generateMetadata()` in the
`page.tsx` files. This issue was only in the CI environment. The workaround to this issue is to put
each `generateMetadata()` function into a `layout.tsx` file. This works because `layout.tsx` files
are not unit tested. See this GitHub issue for more information:
https://www.github.com/vercel/next.js/issues/47299