perf(home): move "start" button under main heading in hero section
A change to improve usability of the home page. The button at the top made look more like a banner
than a clickable button. Having the button underneath text makes it look more like a button.
Additionally, reducing the vertical padding of the hero section allows for a main heading that is
longer (which may be the case for some languages).