fix(compose_form): change tab order of submit and back buttons
The submit ("sign transaction") button should have the focus after tabbing from the last field in
the form because the user most likely wants to submit and move forward, not go back. However,
because the back ("choose template") button was before the submit button in the HTML, the back
button had the focus when tabbing from the last field.