build: disable color output of compile-locales script & remove pre* hook scripts for unit testing
Disabling color makes the output is more readable for the compile-locales script because the
terminal color codes are not mixed in. Also, the locales do not need to be compiled before running
the unit tests. The unit tests do not use any locales because the translation functions are stubbed
and mocked.