feat(app): add theme-switcher
With the theme-switcher the user can manually switch between the themes available. The only themes
available at the moment are "Duck" (light mode theme) and "Duck Dark" (dark mode theme). If the user
does not set the theme, the theme is automatically determined based on the users computer system
preferences.

Closes #30