perf: remove `ShowIf` component & replace use of `ShowIf` with boolean expressions
When using the ShowIf component, the code that is to be conditionally shown is still rendered, even
when the component is to be hidden. The usual boolean expressions that do same thing as the ShowIf
component do not render components that are to be hidden, which improves the performance of the code
by reducing unnecessary renders.