fix(app): allow asset manager, freeze, clawback & reserve addresses to be undefined
For the asset configuration/destroy transaction type, an undefined address is not the same as an
empty address. For a transaction to be a asset-destroy transaction, the manager, freeze, clawback
and reserve addresses must all be `undefined` (or not included). If the manager, freeze, clawback
and reserve addresses are all empty strings, it is still a regular asset-configuration transaction
because at least one of those addresses is specified.