refactor(navbar): remove flags from language menu & separate menu item into its own component
It is better for a language menu to not have country flags because it can be controversal. It is
considered a language menu best-practice not to include them.

Separating the langauge link menu item into its own component removes the need to use React Suspense
in the navbar, which caused the loading of the language button to be delayed.