fix(form): reduce minimum widths of fields with a side label
When a field had a side label, the input was stuck at minimum width of about 100px and ignored
`max-width` and `min-width` CSS properties,  which made fields with longer side labels go beyond the
screen. This issue was caused by browsers making `size` of an input 20 characters when the `size`
attribute is not specified. This default `size` only seems to create an issue when an input box is
set to "flex". Setting a `size` resolves the issue.