fix(send_txn): make sure "always clear" setting is loaded before checking for setting value
Because sessionStorage may take a short while to load, the value of the "always clear transaction
data after sending" setting set by the user was often ignored. Checking the setting value (when
determining whether to delete transaction data) was moved to a `useEffect` where the setting can be
reevaluated after it has loaded.