test(send_txn): remove unit test that checks for "waiting" message
The unit test for checking for the "waiting" message was finicky and failed unpredicably because the
"waiting" message may not have lasted long enough for it to be accurately detected in a unit test.
The test can be removed because the check for the "waiting" is part of the manual tests.