chore(yarn): regenerate lock file
Deleting the old yarn.lock and replacing it with a newly generated one forces Yarn to resolve deep
dependencies again, which causes Yarn to get the newest version of the deep dependencies and remove
some unused deep dependencies. Fortunately, regenerating the yarn.lock file removed the security
vulnerability in the `ip` package, which seems to be an unused deep dependency.