fix(navbar): prevent hydration errors in node selector component by breaking it up
The `<Suspense>` tag that surrounded the node-selector component in the nav-bar was causing
hydration errors. To fix the hydration errors, the node-selector component needed to be restructured
in a way that removed the need the `<Suspense>` wrapper in the nav-bar component. The parts of the
node-selector component that need access to the URL query or path parameters had to be separated
into their own components so those parts would only be used on the client side and could be loaded
dynamically. This is a similar the solutions to the one used when the language-selector, the
compose-transaction-form, and the settings components had a similar hydration errors in the past.