perf(usability): clarify that there is no affiliation with AF or Algo Tech
To make it clear that this project is not affiliated with the Algorand Foundation (AF) or Algorand
Technologies (Algo Tech), a statement saying so is put into the footer. It is possible for a user
new to Algorand to assume this project is affiliated with AF or Algo Tech when various members of
the community recommend using this project.